+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `UniqueMmioPointer::take` for arrays and slices, similar to `get` but taking ownership of
   the original pointer rather than borrowing it.
 - Added implementation of `From` to convert from pointer to array to array of pointers.
+- Implemented `Copy` for `SharedMmioPointer`.
 
 ## 0.2.4